pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mysca"
dynamic = ["version"]
requires-python = ">= 3.8"
readme = "README.md"
license = { file = "LICENSE.txt" }
description = "An implementation of the SCA pipeline."
authors = [
  {name = "Addison Howe"},
]
dependencies = [
  "numpy",
  "scipy",
  "scikit-learn",
  "pandas",
  "matplotlib",
  "seaborn",
  "biopython",
  "jax",
  "imageio",
  "tqdm",
  "ipykernel",
  "ipywidgets",
]

[project.scripts]
sca-preprocess = "mysca.__main__:run_preprocessing"
sca-core = "mysca.__main__:run_sca"
sca-pymol = "mysca.__main__:run_pymol"

[project.optional-dependencies]
dev = [
  "pytest",
]
pymol = [
  "pymol-open-source",
]
